change schema place
